Refactor: Only join string parts if there are any
This refactoring might have a performance impact (cannot measure).
With this patch, the return value in the `filter()` function is not a
`String` anymore, but a list of references: `Vec<&str>`.

Previously, the `String` was only processed further if it was non-empty.
An empty Vec<&str> is the same as an empty `String`: Not relevant for
further processing.

Hence, we can optimize this and join the `Vec<&str>` to `String` only if
it is not empty.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>